examples/requirements.txt
websockets>=11.0.3
redis>=5.0.1
google-generativeai>=0.3.0
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
